pyproject.toml
[tool.poetry]
name = "git-squash-blame"
version = "0.1.0"
description = "Squash git history while preserving blame"
license = "MIT"
authors = ["Evan Grim <evan@mirgnave.com>"]
readme = "README.md"
homepage = "https://github.com/egrim/git-squash-blame"
repository = "https://github.com/egrim/git-squash-blame"
documentation = "https://github.com/egrim/git-squash-blame"
packages = [
    { include = "squashblame.py" }
]

[tool.poetry.scripts]
git-squash-blame = 'squashblame:squash_blame'

[tool.poetry.dependencies]
python = "^3.6"
gitpython = "^3.0.3"
click = "^7.0"

[tool.poetry.dev-dependencies]
ipython = "^7.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
